// ui/src/modules/Circles/List/HealthVariation/interfaces.ts
/*
 * Copyright 2020 ZUP IT SERVICOS EM TECNOLOGIA E INOVACAO SA
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export interface CircleHealth {
  errors: CircleHealthType;
  latency: CircleHealthType;
  requests: CircleRequestsHealth;
}

export interface CircleHealthType {
  circleComponents: [CircleComponentHealth];
  unit: string;
}

export type ComponentHealthStatus = 'STABLE' | 'WARNING' | 'ERROR';

export interface CircleComponentHealth {
  name: string;
  status: ComponentHealthStatus;
  threshold: number;
  value: number;
}

export interface CircleRequestsHealth {
  unit: string;
  value: string;
}